api v1

+++ src/main/java/com/example/mealapp/Entities/Food.java
@@ -5,9 +5,9 @@ public abstract class Food {
     private Long id;
     private String name;
     private int kcal = 0;
-    private int carbs = 0;
-    private int fat = 0;
-    private int protein = 0;
+    private double carbs = 0;
+    private double fat = 0;
+    private double protein = 0;
 
     public void updateData() {
